workarounds: add app-menu and button-layout workarounds
fix-app-menu: if enabled gnome-flashback will set
Gtk/ShellShowsAppMenu to FALSE.

fix-button-layout: if enabled gnome-flashback will set
Gtk/DecorationLayout to value set in this setting. To disable this
workaround set it to empty string.

NOTE: It is not possible to use Gtk/ShellShowsAppMenu and
Gtk/DecorationLayout as overrides in xsettings plugin in
gnome-settings-daemon. gnome-flashback will override these
properties.