notifications: fix crash when removing monitor
When bubble is added to stack we connect to destroy signal to
remove bubble from stack when it is destroyed. When we move same
bubble to other stack old signal handler remains connected and
we end with destroy callback using already freed stack.

Use g_signal_connect_object to make sure that signal handler is
disconnected when stack is destroyed.