input-sources: do not grab keyboard with set-content-type
When gnome-flashback receives the signal of 'set-content-type' from ibus,
gnome-flashback calls gf_keyboard_manager_grab() and gf_keyboard_manager_ungrab()
and the functions change the current input focus in GNOME Xorg and it could
result in closing a popup window which has a password entry by focusing on
the entry.

The solution is to stop to call the APIs on 'set-content-type' signal.

Based on gnome-shell commit:
https://gitlab.gnome.org/GNOME/gnome-shell/commit/551e827841626cd8

https://gitlab.gnome.org/GNOME/gnome-flashback/issues/15