crtc: move configured state to separate struct
To make it more reliable to distinguish between values that are
read from the backend implementation (which is likely to be
irrelevant for anything but the backend implementation), split out
those values (e.g. layout).

This changes the meaning of what was MetaCrtc::rect, to a
MetaCrtcConfig::layout which is the layout the CRTC has in the
global coordinate space.

Based on mutter commits:
https://gitlab.gnome.org/GNOME/mutter/-/commit/fe42d56db39f994144e
https://gitlab.gnome.org/GNOME/mutter/-/commit/9fa56176fd5c95b609c
https://gitlab.gnome.org/GNOME/mutter/-/commit/01aaced1293f360c81c
https://gitlab.gnome.org/GNOME/mutter/-/commit/d2a12ee0fabff9a7f0c