screensaver: queue key events while unlock dialog is not ready
Unlock dialog is created in idle after keyboard and/or mouse
activity and its visibility/readiness depends on PAM.

This means that first key event will never reach unlock dialog if
it was not already created/ready. If user is fast enough more key
events might not reach dialog.

Queue key events in this case and forward to password entry when
dialog becomes visible.

https://gitlab.gnome.org/GNOME/gnome-flashback/-/issues/49