backends: check panel autorotation state before applying autorotation
So far, we've expected this signal to not happen whenever autorotation
shouldn't apply (no accelerometer is a strong reason). In future commits
we'll add further checks to this policy, so prevent autorotation to
change the display configuration if the MetaOrientationManager signal
happens but it should be ignored.

Based on mutter commit:
https://gitlab.gnome.org/GNOME/mutter/-/commit/bb0712077a384400f694