output-xrandr: fix init tile info guard
Monitor tile info is possible to fetch when RANDR version 15 is
exposed by the X11 server. We had inverted the check meaning that
only if older versions were advertised would we attempt to init
the tile information. Fix this guard, thus fix monitor tiling on
X11.

Based on mutter commit:
https://gitlab.gnome.org/GNOME/mutter/-/commit/4ecc80fd802584245