monitor-config-store: fix incorrect string comparison with empty string
strncmp() always return 0 if the passed length is 0. What this means is
that whatever the first string check happens to be, if the parsed XML
cdata was empty (e.g. if we got <element></element>), the first
condition would evaluate to true, which is rather unexpected.

Fix this by making sure the string length is correct first. Also move it
into a helper so we don't need to repeat the same strlen() check every
time.

Based on mutter commit:
https://gitlab.gnome.org/GNOME/mutter/-/commit/6c4380ed417782594f790de246648b56e118420b