autostart: add launcher for Geoclue Demo agent
Geoclue client applications need a Geoclue agent to work, which is autostarted with the session. Gnome-shell contains an internal Geoclue agent, while the Geoclue package provides a demo agent with an autostart file with NotShowIn=GNOME. This causes that the agent is available in all desktop environments except in GNOME Flashback.

This commit adds an autostart file for the demo agend just for GNOME Flashback.

Alternative solution would be to implement a Geoclue agent within GNOME Flashback.